Add a setting for workspaces that controls what LayerGroups are included as part of the workspace
The new setting allows all layergroups to be listed when listing the capabilities when getting the capabilities from a virtual workspace.  The options are: ALL, NONE, AT_LEAST_ONE_LAYER_CONTAINED (at least one layer in the layer groups must be part of the workspace) or ALL_CONTAINED (All layers in workgroups must be contained within the workspace).

The setting is a workspace level setting allowing different workspaces to have different behaviour